+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### Breaking
 ### Features
 - Add `#[ts(crate = "..")]` to allow usage of `#[derive(TS)]` from other proc-macro crates  ([#274](https://github.com/Aleph-Alpha/ts-rs/pull/274))
+- Add support types from `serde_json` behind cargo feature `serde-json-impl`
+
 ### Fixes
 
 # v8.0.0